refatoração de tipos no código
tipos agora se encontram no mesmo arquivo .TS e agora estão seguindo sintaxes mais avançadas do TypeScript
