+++ src/components/Select/index.tsx
@@ -8,20 +8,22 @@ import {
 } from "@/components/ui/select"
 import { Controller } from 'react-hook-form'
 
-export const ShadSelect = ({control, children}: {
+export const ShadSelect = ({control, children, label, name}: {
     control: any,
-    children: React.ReactNode
+    children: React.ReactNode,
+    label?: string,
+    name: string
 }) => {
     return (
         <Controller
-            name='category'
+            name={name}
             control={control}
             defaultValue=''
             rules={{ required: true }}
             render={({ field:{ onChange, value} }) => (
                 <Select onValueChange={onChange} value={value}>
                     <SelectTrigger className="w-full text-paragraph rounded border border-gray-400 px-3 py-2 h-8">
-                        <SelectValue placeholder="Escolha a categoria..." />
+                        <SelectValue placeholder={label || ''} />
                     </SelectTrigger>
                     <SelectContent>
                         {children}
