--- src/types/editItem.ts
@@ -1,5 +1,0 @@
-export interface IEditItemProps {
-    name: string;
-    quantity: number;
-    value: string;
-}
