--- src/types/formItem.ts
@@ -1,5 +1,0 @@
-export interface IFormItem {
-    name: string;
-    category: string;
-    quantity: number;
-}
