--- src/types/product.ts
@@ -1,8 +1,0 @@
-export interface IProductProps {
-    id: string;
-    name: string;
-    category: string;
-    quantity: number;
-    value: string;
-    checked: boolean;
-}
